+++ src/lib.rs
@@ -15,6 +15,10 @@ pub mod flecs_registry;
 
 #[cfg(feature = "arrow_rs")]
 pub mod arrow_snapshot;
+
+#[cfg(feature = "arrow_rs")]
+pub use zip;
+
 pub mod prelude {
     pub use crate::aurora_archive::*;
     pub use crate::bevy_registry::*;